+++ tavern/_core/starlark/starlark_env.py
@@ -2,7 +2,11 @@
 
 import copy
 import dataclasses
+import functools
+import importlib.resources
 import logging
+import re
+import time
 from typing import Any, TypedDict
 
 import starlark
@@ -14,12 +18,25 @@ from tavern._core.run import _TestRunner
 from tavern._core.strict_util import StrictLevel
 from tavern._core.tincture import get_stage_tinctures
 
-from .builtins import get_helpers_source, register_library_builtins, wrap_callable
 from .response_struct import create_response_struct
 from .stage_registry import StageRegistry
 from .types import from_starlark, to_starlark
 
 logger: logging.Logger = logging.getLogger(__name__)
+
+
+def _wrap_callable(fn):
+    """Decorator that converts all arguments from starlark→Python before
+    calling *fn*, and converts the return value from Python→starlark."""
+
+    @functools.wraps(fn)
+    def wrapper(*args, **kwargs):
+        converted_args = [from_starlark(a) for a in args]
+        converted_kwargs = {k: from_starlark(v) for k, v in kwargs.items()}
+        result = fn(*converted_args, **converted_kwargs)
+        return to_starlark(result)
+
+    return wrapper
 
 
 class PipelineContext(TypedDict):
@@ -69,6 +86,19 @@ class StageResponse:
             request_vars=from_starlark(obj["request_vars"]),
             stage_name=obj["stage_name"],
         )
+
+
+def _get_starlark_builtins() -> str:
+    """Load the Starlark builtins from the tavern_helpers.star file.
+
+    Returns:
+        The Starlark code for built-in helper functions
+    """
+    return (
+        importlib.resources.files(__package__)
+        .joinpath("tavern_helpers.star")
+        .read_text()
+    )
 
 
 class StarlarkPipelineRunner:
@@ -133,7 +163,9 @@ class StarlarkPipelineRunner:
         def load(filename: str) -> starlark.FrozenModule:
             """Implements the 'load' function in starlark. Currently only supports loading tavern helpers."""
             if filename == "@tavern_helpers.star":
-                ast = starlark.parse(filename, get_helpers_source(), dialect=dialect)
+                ast = starlark.parse(
+                    filename, _get_starlark_builtins(), dialect=dialect
+                )
                 mod = starlark.Module()
                 self._setup_builtins(mod)
                 starlark.eval(mod, ast, self.globals)
@@ -257,22 +289,19 @@ class StarlarkPipelineRunner:
 
             re = struct(match=_re_match, sub=_re_sub)
 
-        2. Add a wrapper function into builtins.register_library_builtins and add it
-           with module.add_callable. dunder names are used to 'hide' the original
-           function from the user.
+        2. Add a wrapper function into this function and add it with module.add_callable.
+           dunder names are used to 'hide' the original function from the user.
 
-            @wrap_callable
+            @_wrap_callable
             def re_match(pattern, s):
                 return re.match(pattern, s)
 
-            @wrap_callable
+            @_wrap_callable
             def re_sub(pattern, repl, s):
                 return re.sub(pattern, repl, s)
 
             module.add_callable("__re_match", re_match)
             module.add_callable("__re_sub", re_sub)
-
-           Anything registered there is also available in the per-stage expressions.
 
         3. Use from starlark by loading as before:
 
@@ -286,9 +315,7 @@ class StarlarkPipelineRunner:
         for stage_id, stage in self._stage_registry.get_all_stages().items():
             module[stage_id] = to_starlark(stage)
 
-        register_library_builtins(module)
-
-        @wrap_callable
+        @_wrap_callable
         def run_stage_binding(
             stage_id: str, continue_on_fail: bool, extra_vars: dict | None
         ) -> Any:
@@ -310,3 +337,56 @@ class StarlarkPipelineRunner:
                 ) from e
 
         module.add_callable("__run_stage", run_stage_binding)
+
+        @_wrap_callable
+        def log(s: str) -> None:
+            """log a string to stdout."""
+            logger.info(s)
+
+        module.add_callable("log", log)
+
+        @_wrap_callable
+        def re_match(pattern: str, string: str | bytes) -> dict | None:
+            if isinstance(string, bytes):
+                string = string.decode("utf-8")
+            result = re.match(pattern, string)
+            if result is None:
+                return None
+            return {
+                "group0": result.group(0),
+                "groups": list(result.groups()),
+                "start": result.start(),
+                "end": result.end(),
+            }
+
+        module.add_callable("__re_match", re_match)
+
+        @_wrap_callable
+        def re_search(pattern: str, string: str | bytes) -> dict | None:
+            if isinstance(string, bytes):
+                string = string.decode("utf-8")
+            result = re.search(pattern, string)
+            if result is None:
+                return None
+            return {
+                "group0": result.group(0),
+                "groups": list(result.groups()),
+                "start": result.start(),
+                "end": result.end(),
+            }
+
+        module.add_callable("__re_search", re_search)
+
+        @_wrap_callable
+        def re_sub(pattern: str, repl: str, string: str | bytes) -> str:
+            if isinstance(string, bytes):
+                return re.sub(pattern, repl, string.decode("utf-8"))
+            return re.sub(pattern, repl, string)
+
+        module.add_callable("__re_sub", re_sub)
+
+        @_wrap_callable
+        def time_sleep(seconds: float) -> None:
+            time.sleep(seconds)
+
+        module.add_callable("__time_sleep", time_sleep)
